pyproject.toml
[project]
name = "riawelc"
version = "0.2.0"
description = "Welding defect classification and weakly-supervised segmentation pipeline using radiographic images"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Tobias Hoffmann", email = "thoffmann-ml@proton.me"},
]

dependencies = [
    "tensorflow>=2.16,<3",
    "numpy>=1.26",
    "pandas>=2.2",
    "scikit-learn>=1.5",
    "matplotlib>=3.9",
    "seaborn>=0.13",
    "Pillow>=10.4",
    "pyyaml>=6.0",
    "mlflow>=2.16",
    "tqdm>=4.66",
    "structlog>=24.4",
    "jupyterlab>=4.5.4",
    "gradio>=6.5.1",
    "opencv-python>=4.13.0.92",
    "google-cloud-aiplatform>=1.137.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
]
export = [
    "tf2onnx>=1.16",
    "onnxruntime>=1.19",
]
notebooks = [
    "jupyterlab>=4.2",
    "ipywidgets>=8.1",
]
demo = [
    "gradio>=4.44",
]
vertex = [
    "google-cloud-aiplatform>=1.74",
    "google-cloud-storage>=2.18",
]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.6",
    "mypy>=1.11",
]

[dependency-groups]
dev = ["riawelc[dev]"]
api = ["riawelc[api]"]
export = ["riawelc[export]"]
notebooks = ["riawelc[notebooks]"]
demo = ["riawelc[demo]"]
vertex = ["riawelc[vertex]"]
all = ["riawelc[api,export,notebooks,demo,dev]"]

[tool.uv]
conflicts = [
    [
        { extra = "vertex" },
        { extra = "export" },
    ],
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/riawelc"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
